requirements.txt
altgraph==0.17.4
colorama==0.4.6
contourpy==1.3.1
control==0.10.1
CTkMenuBar==0.8
customtkinter==5.2.2
cx_Freeze==7.2.7
cx_Logging==3.2.1
cycler==0.12.1
darkdetect==0.8.0
decorator==5.2.1
et_xmlfile==2.0.0
fonttools==4.55.3
imageio==2.37.0
imageio-ffmpeg==0.6.0
joblib==1.4.2
kiwisolver==1.4.7
lief==0.15.1
markdown-it-py==3.0.0
matplotlib==3.10.0
mdurl==0.1.2
moviepy==2.2.1
mpmath==1.3.0
Nuitka==2.5.7
numpy==2.1.3
openpyxl==3.1.5
ordered-set==4.1.0
packaging==24.2
pandas==2.2.3
pefile==2023.2.7
pillow==11.0.0
proglog==0.1.12
Pygments==2.19.1
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2024.10
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2024.2
pywin32-ctypes==0.2.3
rich==13.9.4
scikit-learn==1.6.0
scipy==1.14.1
setuptools==75.6.0
six==1.17.0
sympy==1.13.3
threadpoolctl==3.5.0
tqdm==4.67.1
tzdata==2024.2
XlsxWriter==3.2.2
zstandard==0.23.0
